test/realcluster: add safe_point case to etcd key (#9112)
ref tikv/pd#8582
