+++ tests/integrations/realcluster/util.go
@@ -15,12 +15,20 @@
 package realcluster
 
 import (
+	"context"
 	"os"
 	"os/exec"
+	"strings"
 	"syscall"
 	"time"
 
+	"github.com/stretchr/testify/require"
+
 	"github.com/pingcap/log"
+
+	pd "github.com/tikv/pd/client"
+	"github.com/tikv/pd/client/opt"
+	"github.com/tikv/pd/client/pkg/caller"
 )
 
 const physicalShiftBits = 18
@@ -58,4 +66,42 @@ func isProcessRunning(pid int) bool {
 	}
 	err = process.Signal(syscall.Signal(0))
 	return err == nil
+}
+
+func newPDClient(re *require.Assertions) pd.Client {
+	pdEndpoints := getPDEndpoints(re)
+	cli, err := pd.NewClientWithContext(
+		context.Background(), caller.TestComponent, pdEndpoints,
+		pd.SecurityOption{}, opt.WithMaxErrorRetry(1),
+	)
+	re.NoError(err)
+	return cli
+}
+
+func getPDEndpoints(re *require.Assertions) []string {
+	output, err := runCommandWithOutput("ps -ef | grep tikv-server | awk -F '--pd-endpoints=' '{print $2}' | awk '{print $1}'")
+	re.NoError(err)
+	var pdAddrs []string
+	for _, addr := range strings.Split(strings.TrimSpace(output), "\n") {
+		// length of addr is less than 5 means it must not be a valid address
+		if len(addr) < 5 {
+			continue
+		}
+		pdAddrs = append(pdAddrs, strings.Split(addr, ",")...)
+	}
+	return removeDuplicates(pdAddrs)
+}
+
+func removeDuplicates(arr []string) []string {
+	uniqueMap := make(map[string]bool)
+	var result []string
+
+	for _, item := range arr {
+		if _, exists := uniqueMap[item]; !exists {
+			uniqueMap[item] = true
+			result = append(result, item)
+		}
+	}
+
+	return result
 }
